+++ test/server/src/language.test.ts
@@ -1,15 +1,13 @@
 import { expect } from 'chai';
 import 'mocha';
-import { Substitute, SubstituteOf, Arg } from '@fluffy-spoon/substitute';
-import { Location, TextDocument } from 'vscode-languageserver';
 
-import { TokenizeMultireplace } from '../../../server/src/language';
+import { tokenizeMultireplace } from '../../../server/src/language';
 
 describe("Multireplace Tokenization", () => {
 	it("should extract a bare variable test", () => {
 		let text = "variable yes | no} extra content";
 
-		let tokens = TokenizeMultireplace(text);
+		let tokens = tokenizeMultireplace(text);
 
 		expect(tokens.test.text).to.equal("variable");
 		expect(tokens.test.index).to.equal(0);
@@ -18,7 +16,7 @@ describe("Multireplace Tokenization", () => {
 	it("should extract a parenthesized test", () => {
 		let text = "(var1 + var2) yes | no} extra content";
 
-		let tokens = TokenizeMultireplace(text);
+		let tokens = tokenizeMultireplace(text);
 
 		expect(tokens.test.text).to.equal("var1 + var2");
 		expect(tokens.test.index).to.equal(1);
@@ -27,7 +25,7 @@ describe("Multireplace Tokenization", () => {
 	it("should extract the bodies", () => {
 		let text = "variable yes | no | maybe } extra content";
 
-		let tokens = TokenizeMultireplace(text);
+		let tokens = tokenizeMultireplace(text);
 
 		expect(tokens.body[0].text).to.equal("yes");
 		expect(tokens.body[0].index).to.equal(9);
@@ -40,7 +38,7 @@ describe("Multireplace Tokenization", () => {
 	it("should find the end index", () => {
 		let text = "variable yes | no} extra content";
 
-		let tokens = TokenizeMultireplace(text);
+		let tokens = tokenizeMultireplace(text);
 
 		expect(tokens.endIndex).to.equal(18);
 	});
@@ -48,7 +46,7 @@ describe("Multireplace Tokenization", () => {
 	it("should extract starting at a given index", () => {
 		let text = "other text @{variable yes | no} extra content";
 
-		let tokens = TokenizeMultireplace(text, 13);
+		let tokens = tokenizeMultireplace(text, 13);
 
 		expect(tokens.test.text).to.equal("variable");
 		expect(tokens.body[0].text).to.equal("yes");
@@ -58,7 +56,7 @@ describe("Multireplace Tokenization", () => {
 	it("should return indices relative to the global index", () => {
 		let text = "other text @{variable yes | no} extra content";
 
-		let tokens = TokenizeMultireplace(text, 13);
+		let tokens = tokenizeMultireplace(text, 13);
 
 		expect(tokens.test.index).to.equal(13);
 		expect(tokens.body[0].index).to.equal(22);